Full Assignment Implementation
The FMOD studio project was created using an example template, this example template being used to aid in the implementation of the sounds in-game. The engine and wind both being driven by the kart’s speed, allowing them to pay louder while speed increases. The script they both use was given through a tutorial available on FMOD’s website.

The brake/drift sound effects are triggered when the drift action is triggered. The impact is triggered the same way but the impact is calculated via the difference in speed within a small time frame. The impact and brake/drift are both using a custom script for each to be calculated and triggered.